refactor: move processMessage out of socket
